+++ lib/securial/error/base_securial_error.rb
@@ -1,8 +1,13 @@
 module Securial
   module Error
-    class BaseSecurialError < StandardError
+    class BaseError < StandardError
+      def self.default_message(message = nil)
+        @default_message = message if message
+        @default_message
+      end
+
       def initialize(message = nil)
-        super(message || "An error occurred in Securial")
+        super(message || self.class.default_message || "An error occurred in Securial")
       end
 
       def backtrace; []; end